refactor: fix Storybook config
